+++ docs/src/markdown/about/changelog.md
@@ -2,6 +2,7 @@
 
 ## 8.3
 
+- **FIX**: BetterEm: Fix some complex cases related to "smart" logic. (#1413)
 - **FIX**: Fix issue that occurs when showing only nth line numbers and using `pymdownx-inline`. Lines not showing a
   line number would not render with the proper leading space.